Misc. changes
Fallback contrast to 223;
Delete unused variable 'endPos';
Better text file viewer interface;
Change cursor blink interval to 200 ticks;
Set U8g2 bus clock;
Reduce screen refresh rate while composing message to reduce flickers.

+++ MultitapIM.hpp
@@ -28,7 +28,7 @@ class MultitapIM {
     uint32_t lastMillis;
     int16_t lastKeycode;
     uint8_t lastPos;
-    uint8_t tickCount;
+    uint16_t tickCount;
     char lastChar;
     bool showingCursor;  // !! necessary?
